feat: adding basic metrics

+++ internal/api/monstera/apimodel/admin_dashboard.go
@@ -4,5 +4,5 @@ package apimodel
 type AdminDashboard struct {
 	LocalUsersCount    int64 `json:"local_users_count"`
 	LocalStatusesCount int64 `json:"local_statuses_count"`
-	OpenReportsCount   int   `json:"open_reports_count"`
+	OpenReportsCount   int64 `json:"open_reports_count"`
 }
